Allow form submit when redirecting.
In login use case it is better to submit the form so browsers will ask
users to save theirs password.
That currently performs a double submit: the first one lets use Ajax to
do the first submit while the second one is only allowed through the
form. Obviously it is not a desirable behaviour and must be improved.